made code more modular
