+++ src/controllers/showHome.js
@@ -1,0 +1,5 @@
+function showHome(req, res) {
+  return res.send('Hello World');
+}
+
+export default showHome;
